remove testapi and related http form
- we shall keep gokeyless lib lean. A tester webservice UI shall
  live on its own project/repo by utilizing gokeyless library.
